[CR] Moved common implementation to common proxy Manager
