+++ v2/internal/plugins/common/proxy/manager.go
@@ -1,15 +1,25 @@
 package proxy
 
-import "sync"
+import (
+	"fmt"
+	"sync"
+
+	"github.com/librato/snap-plugin-lib-go/v2/plugin"
+)
 
 type ContextManager struct {
 	activeTasksMutex sync.RWMutex        // mutex associated with activeTasks
 	activeTasks      map[string]struct{} // map of active tasks (tasks for which Collect RPC request is progressing)
+
+	TasksLimit     int
+	InstancesLimit int
 }
 
 func NewContextManager() *ContextManager {
 	return &ContextManager{
-		activeTasks: map[string]struct{}{},
+		activeTasks:    map[string]struct{}{},
+		TasksLimit:     plugin.NoLimit,
+		InstancesLimit: plugin.NoLimit,
 	}
 }
 
@@ -30,4 +40,22 @@ func (cm *ContextManager) MarkTaskAsCompleted(id string) {
 	defer cm.activeTasksMutex.Unlock()
 
 	delete(cm.activeTasks, id)
+}
+
+func (cm *ContextManager) DefineTasksPerInstanceLimit(limit int) error {
+	if limit < -1 {
+		return fmt.Errorf("invalid tasks limit")
+	}
+
+	cm.TasksLimit = limit
+	return nil
+}
+
+func (cm *ContextManager) DefineInstancesLimit(limit int) error {
+	if limit < -1 {
+		return fmt.Errorf("invalid instances limit")
+	}
+
+	cm.InstancesLimit = limit
+	return nil
 }
